Refactor endpoint validation source helpers
